kallsyms: Stop considering the gate area to be kernel text
The gate area isn't kernel text.  It's user legacy garbage.  The
check is also nonsense -- whether the vsyscall page on x86 is valid
at all depends on the mm.

This will result in vsprintf being unwilling to try to decode gate
area addresses.  I'd be rather surprised if it ever worked in the
first place, though, given that the gate area is just an alias and
isn't likely to show up in kallsyms in the first place.  (I checked.
It doesn't on x86_64.  The only other architectures that even have
gate areas are arm and ia64.)

This gets rid of the only user of in_gate_area_no_mm().

Signed-off-by: Andy Lutomirski <luto@kernel.org>