x86/ptrace: Fix xfpregs_set()'s incorrect xmm clearing
xfpregs_set() handles 32-bit REGSET_XFP and 64-bit REGSET_FP.  The actual
code treats these regsets as modern FX state (i.e. the beginning part of
XSTATE).  The declarations of the regsets thought they were the legacy i387
format.  The code thought they were the 32-bit (no xmm8..15) variant of
XSTATE and, for good measure, made the high bits disappear by zeroing the
wrong part of the buffer.  The latter broke ptrace, and everything else
confused anyone trying to understand the code.  In particular, the nonsense
definitions of the regsets confused me when I wrote this code.

Clean this all up.  Change the declarations to match reality (which
shouldn't change the generated code, let alone the ABI) and fix
xfpregs_set() to clear the correct bits and to only do so for 32-bit
callers.

[completely untested]

Link: https://bugzilla.kernel.org/show_bug.cgi?id=215524
Fixes: 6164331d15f7 ("x86/fpu: Rewrite xfpregs_set()")
Signed-off-by: Andy Lutomirski <luto@kernel.org>