net/mlx5e: Fix mqprio_rl handling on devlink reload
Keep mqprio_rl data to params and restore the configuration in case of
devlink reload.
Change the location of mqprio_rl resources cleanup so it will be done
also in reload flow.

Also, remove the rl pointer from the params, since this is dynamic object
and saved to priv.

Issue: 3019878
Signed-off-by: Moshe Tal <moshet@nvidia.com>
Change-Id: I1b6377e753de8ebfffc25a6c5826cc9f15ebbc5c
Reviewed-by: Tariq Toukan <tariqt@nvidia.com>
Signed-off-by: Saeed Mahameed <saeedm@nvidia.com>