afs: Fix dynamic root getattr
The recent patch to make afs_getattr consult the server didn't account for
the pseudo-inodes employed by the dynamic root-type afs superblock not
having a volume or a server to access, and thus an oops occurs if such a
directory is stat'd.

Fix this by checking to see if the vnode->volume pointer actually points
anywhere before following it in afs_getattr().

This can be tested by stat'ing a directory in /afs.  It may be sufficient
just to do "ls /afs" and the oops looks something like:

	BUG: kernel NULL pointer dereference, address: 0000000000000020
	...
	RIP: 0010:afs_getattr+0x8b/0x14b
	...
	Call Trace:
	 <TASK>
	 vfs_statx+0x79/0xf5
	 vfs_fstatat+0x49/0x62

Fixes: 2aeb8c86d499 ("afs: Fix afs_getattr() to refetch file status if callback break occurred")
Reported-by: Marc Dionne <marc.dionne@auristor.com>
Signed-off-by: David Howells <dhowells@redhat.com>
cc: linux-afs@lists.infradead.org